Désengorger le fichier defi.js du cafe de la place en répartissant les datas d'orientation, du parcours haut et du parcours bas en sous fichiers
