+++ src/situations/cafe_de_la_place/data/parcours_haut.js
@@ -1,0 +1,14 @@
+const sousConsigneHPar1 = {
+  id: 'HPar-sous-consigne-1',
+  nom_technique: 'sous_consigne_HPar_1',
+  type: 'sous-consigne',
+  intitule: "Première sous-consigne du parcours Haut à venir",
+};
+
+const parcoursHaut = {
+  series: [
+    { cartes: [ sousConsigneHPar1 ] },
+  ]
+};
+
+export { parcoursHaut };
